requirements.txt
pandas==2.0.2
mysql-connector==2.2.9
python-decouple==3.8